Change the rating from float to postiive integer field

+++ ratings/models.py
@@ -11,7 +11,7 @@ class Rating(models.Model):
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
     content = models.TextField()
-    rating = models.FloatField(
+    rating = models.PositiveIntegerField(
         validators=[MinValueValidator(1), MaxValueValidator(5)]
     )
